+++ src/components/layout/Gnb/index.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import { useRouter } from 'next/router';
 
-import useDropdownStore from '@/stores/useDropdownStore';
+import useDropdownStore from '@/stores/dropdown/useDropdownStore';
 
 import Dropdown from './Dropdown';
